docker/docker_submission/requirements.txt
tables
scikit-image
scipy
numpy
torch
tiffile
xmltodict
scikit-learn
Pillow
matplotlib
pandas

